api/requirements.txt
fastapi[standard]
mangum
getstream
psycopg2-binary
sqlalchemy
python-decouple